did alot here tbh, GX2 rendering is broken currently though

+++ Oxygen/lemonscript/source/lemon/translator/Nativizer.h
@@ -76,7 +76,7 @@ namespace lemon
 
 		struct LookupDictionary
 		{
-			void addEmptyEntries(const uint64* hashes, size_t numHashes);
+			void addEmptyEntries(const uint8* hashes, size_t numHashes);
 			void loadFunctions(const CompactFunctionEntry* entries, size_t numEntries);
 			void loadParameterInfo(const uint8* data, size_t count);
